move plugins to folder
